Init PackagerRequest and WriteBundlesRequest
