+++ packages/core/core/src/PackagerRunner.js
@@ -38,10 +38,7 @@ import {serialize} from './serializer';
 
 type Opts = {|
   config: ParcelConfig,
-  configRef?: SharedReference,
-  farm?: WorkerFarm,
   options: ParcelOptions,
-  optionsRef?: SharedReference,
   report: ReportFn,
 |};
 
@@ -64,132 +61,17 @@ const BOUNDARY_LENGTH = HASH_REF_PREFIX.length + 32 - 1;
 
 export default class PackagerRunner {
   config: ParcelConfig;
-  configRef: ?SharedReference;
   options: ParcelOptions;
-  optionsRef: ?SharedReference;
-  farm: ?WorkerFarm;
   pluginOptions: PluginOptions;
   distDir: FilePath;
   distExists: Set<FilePath>;
   report: ReportFn;
-  getBundleInfoFromWorker: ({|
-    bundle: InternalBundle,
-    bundleGraphReference: SharedReference,
-    configRef: SharedReference,
-    optionsRef: SharedReference,
-  |}) => Promise<BundleInfo>;
 
-  constructor({config, configRef, farm, options, optionsRef, report}: Opts) {
+  constructor({config, options, report}: Opts) {
     this.config = config;
-    this.configRef = configRef;
     this.options = options;
-    this.optionsRef = optionsRef;
     this.pluginOptions = new PluginOptions(this.options);
-
-    this.farm = farm;
     this.report = report;
-    this.getBundleInfoFromWorker = farm
-      ? farm.createHandle('runPackage')
-      : () => {
-          throw new Error(
-            'Cannot call PackagerRunner.writeBundleFromWorker() in a worker',
-          );
-        };
-  }
-
-  async writeBundles(bundleGraph: InternalBundleGraph) {
-    let farm = nullthrows(this.farm);
-    let {ref, dispose} = await farm.createSharedReference(
-      bundleGraph,
-      serialize(bundleGraph),
-    );
-
-    let bundleInfoMap: {|
-      [string]: {|
-        ...BundleInfo,
-        cacheKeys: CacheKeyMap,
-      |},
-    |} = {};
-    let writeEarlyPromises = {};
-    let hashRefToNameHash = new Map();
-    let bundles = bundleGraph.getBundles().filter(bundle => {
-      // Do not package and write placeholder bundles to disk. We just
-      // need to update the name so other bundles can reference it.
-      if (bundle.isPlaceholder) {
-        let hash = bundle.id.slice(-8);
-        hashRefToNameHash.set(bundle.hashReference, hash);
-        let name = nullthrows(bundle.name).replace(bundle.hashReference, hash);
-        bundle.filePath = path.join(bundle.target.distDir, name);
-        return false;
-      }
-
-      // skip inline bundles, they will be processed via the parent bundle
-      return !bundle.isInline;
-    });
-
-    try {
-      await Promise.all(
-        bundles.map(async bundle => {
-          let info = await this.processBundle(bundle, bundleGraph, ref);
-          bundleInfoMap[bundle.id] = info;
-          if (!info.hashReferences.length) {
-            hashRefToNameHash.set(
-              bundle.hashReference,
-              this.options.shouldContentHash
-                ? info.hash.slice(-8)
-                : bundle.id.slice(-8),
-            );
-            writeEarlyPromises[bundle.id] = this.writeToDist({
-              bundle,
-              info,
-              hashRefToNameHash,
-              bundleGraph,
-            });
-          }
-        }),
-      );
-      assignComplexNameHashes(
-        hashRefToNameHash,
-        bundles,
-        bundleInfoMap,
-        this.options,
-      );
-      await Promise.all(
-        bundles.map(
-          bundle =>
-            writeEarlyPromises[bundle.id] ??
-            this.writeToDist({
-              bundle,
-              info: bundleInfoMap[bundle.id],
-              hashRefToNameHash,
-              bundleGraph,
-            }),
-        ),
-      );
-    } finally {
-      await dispose();
-    }
-  }
-
-  async processBundle(
-    bundle: InternalBundle,
-    bundleGraph: InternalBundleGraph,
-    bundleGraphReference: SharedReference,
-  ): Promise<{|
-    ...BundleInfo,
-    cacheKeys: CacheKeyMap,
-  |}> {
-    let start = Date.now();
-
-    return {
-      ...(await this.getBundleInfoFromWorker({
-        bundle,
-        bundleGraphReference,
-        optionsRef: nullthrows(this.optionsRef),
-        configRef: nullthrows(this.configRef),
-      })),
-      time: Date.now() - start,
-    };
   }
 
   async loadConfigs(
@@ -531,79 +413,6 @@ export default class PackagerRunner {
     };
   }
 
-  async writeToDist({
-    bundle,
-    bundleGraph,
-    info,
-    hashRefToNameHash,
-  }: {|
-    bundle: InternalBundle,
-    bundleGraph: InternalBundleGraph,
-    info: {|...BundleInfo, cacheKeys: CacheKeyMap|},
-    hashRefToNameHash: Map<string, string>,
-  |}) {
-    let {inputFS, outputFS} = this.options;
-    let name = nullthrows(bundle.name);
-    let thisHashReference = bundle.hashReference;
-
-    if (info.type !== bundle.type) {
-      name = name.slice(0, -path.extname(name).length) + '.' + info.type;
-      bundle.type = info.type;
-    }
-
-    if (name.includes(thisHashReference)) {
-      let thisNameHash = nullthrows(hashRefToNameHash.get(thisHashReference));
-      name = name.replace(thisHashReference, thisNameHash);
-    }
-
-    let filePath = path.join(bundle.target.distDir, name);
-    bundle.filePath = filePath;
-
-    let dir = path.dirname(filePath);
-    await outputFS.mkdirp(dir); // ? Got rid of dist exists, is this an expensive operation
-
-    // Use the file mode from the entry asset as the file mode for the bundle.
-    // Don't do this for browser builds, as the executable bit in particular is unnecessary.
-    let publicBundle = NamedBundle.get(bundle, bundleGraph, this.options);
-    let mainEntry = publicBundle.getMainEntry();
-    let writeOptions =
-      publicBundle.env.isBrowser() || !mainEntry
-        ? undefined
-        : {
-            mode: (await inputFS.stat(mainEntry.filePath)).mode,
-          };
-    let cacheKeys = info.cacheKeys;
-    let contentStream = this.options.cache.getStream(cacheKeys.content);
-    let size = await writeFileStream(
-      outputFS,
-      filePath,
-      contentStream,
-      info.hashReferences,
-      hashRefToNameHash,
-      writeOptions,
-    );
-    bundle.stats = {
-      size,
-      time: info.time ?? 0,
-    };
-
-    let mapKey = cacheKeys.map;
-    if (
-      bundle.env.sourceMap &&
-      !bundle.env.sourceMap.inline &&
-      (await this.options.cache.has(mapKey))
-    ) {
-      let mapStream = this.options.cache.getStream(mapKey);
-      await writeFileStream(
-        outputFS,
-        filePath + '.map',
-        mapStream,
-        info.hashReferences,
-        hashRefToNameHash,
-      );
-    }
-  }
-
   async writeToCache(
     cacheKeys: CacheKeyMap,
     type: string,
@@ -670,99 +479,4 @@ export default class PackagerRunner {
   static getInfoKey(cacheKey: string): string {
     return hashString(`${cacheKey}:info`);
   }
-}
-
-function writeFileStream(
-  fs: FileSystem,
-  filePath: FilePath,
-  stream: Readable,
-  hashReferences: Array<string>,
-  hashRefToNameHash: Map<string, string>,
-  options: ?FileOptions,
-): Promise<number> {
-  return new Promise((resolve, reject) => {
-    let initialStream = hashReferences.length
-      ? stream.pipe(replaceStream(hashRefToNameHash))
-      : stream;
-    let fsStream = fs.createWriteStream(filePath, options);
-    let fsStreamClosed = new Promise(resolve => {
-      fsStream.on('close', () => resolve());
-    });
-    let bytesWritten = 0;
-    initialStream
-      .pipe(
-        new TapStream(buf => {
-          bytesWritten += buf.length;
-        }),
-      )
-      .pipe(fsStream)
-      .on('finish', () => resolve(fsStreamClosed.then(() => bytesWritten)))
-      .on('error', reject);
-  });
-}
-
-function replaceStream(hashRefToNameHash) {
-  let boundaryStr = '';
-  return new Transform({
-    transform(chunk, encoding, cb) {
-      let str = boundaryStr + chunk.toString();
-      let replaced = str.replace(HASH_REF_REGEX, match => {
-        return hashRefToNameHash.get(match) || match;
-      });
-      boundaryStr = replaced.slice(replaced.length - BOUNDARY_LENGTH);
-      let strUpToBoundary = replaced.slice(
-        0,
-        replaced.length - BOUNDARY_LENGTH,
-      );
-      cb(null, strUpToBoundary);
-    },
-
-    flush(cb) {
-      cb(null, boundaryStr);
-    },
-  });
-}
-
-function assignComplexNameHashes(
-  hashRefToNameHash,
-  bundles,
-  bundleInfoMap,
-  options,
-) {
-  for (let bundle of bundles) {
-    if (hashRefToNameHash.get(bundle.hashReference) != null) {
-      continue;
-    }
-
-    hashRefToNameHash.set(
-      bundle.hashReference,
-      options.shouldContentHash
-        ? hashString(
-            [...getBundlesIncludedInHash(bundle.id, bundleInfoMap)]
-              .map(bundleId => bundleInfoMap[bundleId].hash)
-              .join(':'),
-          ).slice(-8)
-        : bundle.id.slice(-8),
-    );
-  }
-}
-
-function getBundlesIncludedInHash(
-  bundleId,
-  bundleInfoMap,
-  included = new Set(),
-) {
-  included.add(bundleId);
-  for (let hashRef of bundleInfoMap[bundleId].hashReferences) {
-    let referencedId = getIdFromHashRef(hashRef);
-    if (!included.has(referencedId)) {
-      getBundlesIncludedInHash(referencedId, bundleInfoMap, included);
-    }
-  }
-
-  return included;
-}
-
-function getIdFromHashRef(hashRef: string) {
-  return hashRef.slice(HASH_REF_PREFIX.length);
 }
